requirements.txt
asgiref==3.7.2
Brotli==1.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cloudinary==1.36.0
colorama==0.4.6
coloredlogs==15.0.1
crispy-bootstrap5==0.7
cssselect2==0.7.0
dj-database-url==2.1.0
Django==4.2.4
django-cloudinary-storage==0.3.0
django-cors-headers==4.2.0
django-crispy-forms==2.0
django-environ==0.11.2
django-filter==23.2
django-renderpdf==4.0.0
djangorestframework==3.14.0
flatbuffers==23.5.26
fontawesome-free==5.15.4
fonttools==4.43.1
gunicorn==21.2.0
html5lib==1.1
humanfriendly==10.0
idna==3.4
imageio==2.31.5
lazy_loader==0.3
llvmlite==0.41.0
more-itertools==10.1.0
mpmath==1.3.0
networkx==3.1
numba==0.58.0
numpy==1.25.2
onnxruntime==1.16.0
opencv-python-headless==4.8.1.78
packaging==23.2
Pillow==10.0.0
platformdirs==3.11.0
pooch==1.7.0
protobuf==4.24.4
psycopg2-binary==2.9.9
pycparser==2.21
pydyf==0.8.0
PyMatting==1.1.10
pyphen==0.14.0
pyreadline3==3.4.1
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
rembg==2.0.50
requests==2.31.0
scikit-image==0.22.0
scipy==1.11.3
six==1.16.0
sqlparse==0.4.4
sslcommerz-lib==1.0
sympy==1.12
tifffile==2023.9.26
tinycss2==1.2.1
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.6
weasyprint==60.1
webencodings==0.5.1
whitenoise==6.6.0
zopfli==0.2.3
